feat: props inferred for page components
