+++ packages/waku/src/lib/plugins/vite-plugin-fs-router-typegen.ts
@@ -168,6 +168,7 @@ import type { PathsForPages } from 'waku/router';\n\n`;
   declare module 'waku/router' {
     interface RouteConfig {
       paths: PathsForPages<typeof _pages>;
+      pages: typeof _pages;
     }
   }
