+++ packages/waku/src/router/create-pages.ts
@@ -10,8 +10,10 @@ import {
   path2regexp,
 } from '../lib/utils/path.js';
 import type { PathSpec } from '../lib/utils/path.js';
-import type { Split } from './util-types.js';
-import type { AnyPage } from './create-pages-utils/inferred-path-types.js';
+import type {
+  AnyPage,
+  GetSlugs,
+} from './create-pages-utils/inferred-path-types.js';
 
 const hasPathSpecPrefix = (prefix: PathSpec, path: PathSpec) => {
   for (let i = 0; i < prefix.length; i++) {
@@ -66,20 +68,6 @@ export type PathWithSlug<T, K extends string> =
       ? T
       : never
     : never;
-
-type _GetSlugs<
-  Route extends string,
-  SplitRoute extends string[] = Split<Route, '/'>,
-  Result extends string[] = [],
-> = SplitRoute extends []
-  ? Result
-  : SplitRoute extends [`${infer MaybeSlug}`, ...infer Rest extends string[]]
-    ? MaybeSlug extends `[${infer Slug}]`
-      ? _GetSlugs<Route, Rest, [...Result, Slug]>
-      : _GetSlugs<Route, Rest, Result>
-    : Result;
-
-export type GetSlugs<Route extends string> = _GetSlugs<Route>;
 
 export type StaticSlugRoutePathsTuple<
   T extends string,
